+++ src/test/PlaywrightTest/Path/solicitor/applyCaveat.spec.ts
@@ -1,23 +1,23 @@
 import dateFns from "date-fns";
-import { test } from "../../../Fixtures/fixtures.ts";
+import { test } from "../../Fixtures/fixtures.ts";
 
 // const {runAccessibilityTest} = require('../../Accessibility/axeUtils');
 // const testConfig = require('src/test/config');
-import createCaseConfig from "../../../Pages/createCase/createCaseConfig.json" with { type: "json" };
+import createCaseConfig from "../../Pages/createCase/createCaseConfig.json" with { type: "json" };
 
-import applicationDetailsConfig from "../../../Pages/solicitorApplyCaveat/applicationDetails/applicationDetails.json" with { type: "json" };
-import applyCaveatConfig from "../../../Pages/solicitorApplyCaveat/applyCaveat/applyCaveat.json" with { type: "json" };
-import completeApplicationConfig from "../../../Pages/solicitorApplyCaveat/completeApplication/completeApplication.json" with { type: "json" };
+import applicationDetailsConfig from "../../Pages/solicitorApplyCaveat/applicationDetails/applicationDetails.json" with { type: "json" };
+import applyCaveatConfig from "../../Pages/solicitorApplyCaveat/applyCaveat/applyCaveat.json" with { type: "json" };
+import completeApplicationConfig from "../../Pages/solicitorApplyCaveat/completeApplication/completeApplication.json" with { type: "json" };
 
-import { testConfig } from "../../../Configs/config.ts";
-import caseDetailsTabConfig from "../../../Pages/caseDetails/solicitorApplyCaveat/caseDetailsTabConfig.json" with { type: "json" };
-import caveatDetailsTabConfig from "../../../Pages/caseDetails/solicitorApplyCaveat/caveatDetailsTabConfig.json" with { type: "json" };
-import caveatorDetailsTabConfig from "../../../Pages/caseDetails/solicitorApplyCaveat/caveatorDetailsTabConfig.json" with { type: "json" };
-import deceasedDetailsTabConfig from "../../../Pages/caseDetails/solicitorApplyCaveat/deceasedDetailsTabConfig.json" with { type: "json" };
-import historyTabConfig from "../../../Pages/caseDetails/solicitorApplyCaveat/historyTabConfig.json" with { type: "json" };
-import notificationsTabConfig from "../../../Pages/caseDetails/solicitorApplyCaveat/notificationsTabConfig.json" with { type: "json" };
-import serviceRequestReviewTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/serviceRequestReviewTabConfig.json" with { type: "json" };
-import serviceRequestTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/serviceRequestTabConfig.json" with { type: "json" };
+import { testConfig } from "../../Configs/config.ts";
+import caseDetailsTabConfig from "../../Pages/caseDetails/solicitorApplyCaveat/caseDetailsTabConfig.json" with { type: "json" };
+import caveatDetailsTabConfig from "../../Pages/caseDetails/solicitorApplyCaveat/caveatDetailsTabConfig.json" with { type: "json" };
+import caveatorDetailsTabConfig from "../../Pages/caseDetails/solicitorApplyCaveat/caveatorDetailsTabConfig.json" with { type: "json" };
+import deceasedDetailsTabConfig from "../../Pages/caseDetails/solicitorApplyCaveat/deceasedDetailsTabConfig.json" with { type: "json" };
+import historyTabConfig from "../../Pages/caseDetails/solicitorApplyCaveat/historyTabConfig.json" with { type: "json" };
+import notificationsTabConfig from "../../Pages/caseDetails/solicitorApplyCaveat/notificationsTabConfig.json" with { type: "json" };
+import serviceRequestReviewTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/serviceRequestReviewTabConfig.json" with { type: "json" };
+import serviceRequestTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/serviceRequestTabConfig.json" with { type: "json" };
 
 test.describe("Solicitor - Apply Caveat", () => {
   test("Solicitor - Apply Caveat @webkit", async ({
