+++ src/test/PlaywrightTest/Path/solicitor/applyGrantOfProbateNoWillEE.spec.ts
@@ -1,27 +1,27 @@
-import { test } from "../../../Fixtures/fixtures.ts";
+import { test } from "../../Fixtures/fixtures.ts";
 
-import createCaseConfig from "../../../Pages/createCase/createCaseConfig.json" with { type: "json" };
+import createCaseConfig from "../../Pages/createCase/createCaseConfig.json" with { type: "json" };
 
 // const testConfig = require('src/test/config');
 // const createCaseConfig = require('src/test/end-to-end/pages/createCase/createCaseConfig');
 
-import applyProbateConfig from "../../../Pages/solicitorApplyProbate/applyProbate/applyProbateConfig.json" with { type: "json" };
-import deceasedDetailsConfig from "../../../Pages/solicitorApplyProbate/deceasedDetails/deceasedDetailsConfig.json" with { type: "json" };
-import intestacyDetailsConfig from "../../../Pages/solicitorApplyProbate/intestacyDetails/intestacyDetails.json" with { type: "json" };
-import completeApplicationConfig from "../../../Pages/solicitorApplyProbate/completeApplication/completeApplication.json" with { type: "json" };
+import applyProbateConfig from "../../Pages/solicitorApplyProbate/applyProbate/applyProbateConfig.json" with { type: "json" };
+import deceasedDetailsConfig from "../../Pages/solicitorApplyProbate/deceasedDetails/deceasedDetailsConfig.json" with { type: "json" };
+import intestacyDetailsConfig from "../../Pages/solicitorApplyProbate/intestacyDetails/intestacyDetails.json" with { type: "json" };
+import completeApplicationConfig from "../../Pages/solicitorApplyProbate/completeApplication/completeApplication.json" with { type: "json" };
 
-import applicantDetailsTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/applicantDetailsTabConfig.json" with { type: "json" };
-import deceasedTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/deceasedTabConfigEE.json" with { type: "json" };
-import iHTTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/iHTTabConfigEENoForm.json" with { type: "json" };
-import caseDetailsTabDeceasedDtlsConfig from "../../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabDeceasedDtlsConfigEENo.json" with { type: "json" };
-import caseDetailsTabIntestacyConfig from "../../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabIntestacyConfig.json" with { type: "json" };
-import caseDetailsTabUpdatesConfig from "../../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabUpdatesConfig.json" with { type: "json" };
+import applicantDetailsTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/applicantDetailsTabConfig.json" with { type: "json" };
+import deceasedTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/deceasedTabConfigEE.json" with { type: "json" };
+import iHTTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/iHTTabConfigEENoForm.json" with { type: "json" };
+import caseDetailsTabDeceasedDtlsConfig from "../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabDeceasedDtlsConfigEENo.json" with { type: "json" };
+import caseDetailsTabIntestacyConfig from "../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabIntestacyConfig.json" with { type: "json" };
+import caseDetailsTabUpdatesConfig from "../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabUpdatesConfig.json" with { type: "json" };
 
-import sotTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/sotTabConfig.json" with { type: "json" };
-import copiesTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/copiesTabConfig.json" with { type: "json" };
-import historyTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/historyTabConfig.json" with { type: "json" };
-import serviceRequestTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/serviceRequestTabConfig.json" with { type: "json" };
-import serviceRequestReviewTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/serviceRequestReviewTabConfig.json" with { type: "json" };
+import sotTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/sotTabConfig.json" with { type: "json" };
+import copiesTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/copiesTabConfig.json" with { type: "json" };
+import historyTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/historyTabConfig.json" with { type: "json" };
+import serviceRequestTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/serviceRequestTabConfig.json" with { type: "json" };
+import serviceRequestReviewTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/serviceRequestReviewTabConfig.json" with { type: "json" };
 
 test.describe("Solicitor - Apply Grant of probate Excepted Estates - No Will (Intestacy)", () => {
   test("Solicitor - Apply Grant of probate Excepted Estates @firefox", async ({
