+++ src/test/PlaywrightTest/Path/solicitor/applyGrantOfProbateSingleExecutorEE.spec.ts
@@ -1,32 +1,32 @@
 import dotenv from "dotenv";
 dotenv.config();
 
-import { test } from '../../../Fixtures/index.ts';
-import { testConfig } from "../../../Configs/config.ts";
+import { test } from '../../Fixtures/index.ts';
+import { testConfig } from "../../Configs/config.ts";
 
-import createCaseConfig from "../../../Pages/createCase/createCaseConfig.json" with { type: "json" };
+import createCaseConfig from "../../Pages/createCase/createCaseConfig.json" with { type: "json" };
 
-import applyProbateConfig from "../../../Pages/solicitorApplyProbate/applyProbate/applyProbateConfig.json" with { type: "json" };
-import gopConfig from "../../../Pages/solicitorApplyProbate/grantOfProbate/grantOfProbate.json" with { type: "json" };
-import deceasedDetailsConfig from "../../../Pages/solicitorApplyProbate/deceasedDetails/deceasedDetailsConfig.json" with { type: "json" };
-import completeApplicationConfig from "../../../Pages/solicitorApplyProbate/completeApplication/completeApplication.json" with { type: "json" };
+import applyProbateConfig from "../../Pages/solicitorApplyProbate/applyProbate/applyProbateConfig.json" with { type: "json" };
+import gopConfig from "../../Pages/solicitorApplyProbate/grantOfProbate/grantOfProbate.json" with { type: "json" };
+import deceasedDetailsConfig from "../../Pages/solicitorApplyProbate/deceasedDetails/deceasedDetailsConfig.json" with { type: "json" };
+import completeApplicationConfig from "../../Pages/solicitorApplyProbate/completeApplication/completeApplication.json" with { type: "json" };
 
-import applicantDetailsTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/applicantDetailsTabConfig.json" with { type: "json" };
-import applicantExecutorDetailsTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/applicantExecDetailsTrustCorpTabConfig.json" with { type: "json" };
-import deceasedTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/deceasedTabConfigEE.json" with { type: "json" };
-import iHTTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/iHTTabConfigEE400.json" with { type: "json" };
-import caseDetailsTabDeceasedDtlsConfig from "../../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabDeceasedDtlsConfigEE.json" with { type: "json" };
-import caseDetailsTabGopConfig from "../../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabGopTrustCorpConfig.json" with { type: "json" };
-import caseDetailsTabUpdatesConfig from "../../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabUpdatesConfig.json" with { type: "json" };
-import sotTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/sotTabConfig.json" with { type: "json" };
-import copiesTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/copiesTabConfig.json" with { type: "json" };
-import historyTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/historyTabConfig.json" with { type: "json" };
-import serviceRequestTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/serviceRequestTabConfig.json" with { type: "json" };
-import serviceRequestReviewTabConfig from "../../../Pages/caseDetails/solicitorApplyProbate/serviceRequestReviewTabConfig.json" with { type: "json" };
-import caseProgressConfig from "../../../Pages/caseProgressStandard/caseProgressConfig.json" with { type: "json" };
-import refundConfig from "../../../Pages/solicitorApplyProbate/makePayment/refundConfig.json" with { type: "json" };
-import { refundReviewConfig } from "../../../Pages/solicitorApplyProbate/makePayment/refundReviewConfig.ts";
-import {getAccessToken, getServiceAuthToken} from "../../../Pages/utility/apiHelper.ts";
+import applicantDetailsTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/applicantDetailsTabConfig.json" with { type: "json" };
+import applicantExecutorDetailsTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/applicantExecDetailsTrustCorpTabConfig.json" with { type: "json" };
+import deceasedTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/deceasedTabConfigEE.json" with { type: "json" };
+import iHTTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/iHTTabConfigEE400.json" with { type: "json" };
+import caseDetailsTabDeceasedDtlsConfig from "../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabDeceasedDtlsConfigEE.json" with { type: "json" };
+import caseDetailsTabGopConfig from "../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabGopTrustCorpConfig.json" with { type: "json" };
+import caseDetailsTabUpdatesConfig from "../../Pages/caseDetails/solicitorApplyProbate/caseDetailsTabUpdatesConfig.json" with { type: "json" };
+import sotTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/sotTabConfig.json" with { type: "json" };
+import copiesTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/copiesTabConfig.json" with { type: "json" };
+import historyTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/historyTabConfig.json" with { type: "json" };
+import serviceRequestTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/serviceRequestTabConfig.json" with { type: "json" };
+import serviceRequestReviewTabConfig from "../../Pages/caseDetails/solicitorApplyProbate/serviceRequestReviewTabConfig.json" with { type: "json" };
+import caseProgressConfig from "../../Pages/caseProgressStandard/caseProgressConfig.json" with { type: "json" };
+import refundConfig from "../../Pages/solicitorApplyProbate/makePayment/refundConfig.json" with { type: "json" };
+import { refundReviewConfig } from "../../Pages/solicitorApplyProbate/makePayment/refundReviewConfig.ts";
+import {getAccessToken, getServiceAuthToken} from "../../Pages/utility/apiHelper.ts";
 
 test.describe.serial("Solicitor - Apply Grant of probate Excepted Estates and Refunds", () => {
   let caseRef;
